Update README.md
A JavaScript issue with select options in a scroll-box.
Hello, I am a beginner at programming and, as an exercise for a reservation, I want to make a scroll-box with various options in it to be displayed on my page, alongside the first name and the last name text-fields. The problem is that. When a user, for example, enters 2 in the input text-field for how many persons will be booked, I have 2 text-fields for first name, last name and age but 2 scroll-boxes in a different place (I mean not each one of them right after the age text-nodes where I
want them) and 2 times the options in the same
scroll-box. When I enter a different number in the input text-field, for example 3, I have 3 scroll-boxes plus the 2 from the previous time and 5 times the options in the same scroll-box! I tried to inspect the issue through the inspector of the web-page to see what I have wrong with the structure but I haven't come up with much. If you'd be kind to give me a little help over here, thank you in advance.